+++ examples/test_slack.py
@@ -17,16 +17,17 @@ token = CRED_PATH.read_text().strip()
 @pytest.fixture(scope='module')
 async def exec():
     async with aiohttp.ClientSession() as client:
-        yield slack.async_executor(auth=token, client=client)
+        yield snug.async_executor(auth=slack.token_auth(token),
+                                  client=client)
 
 
 @pytest.mark.asyncio
 async def test_channel_list(exec):
-    lookup = slack.channels.list(exclude_archived=True)
+    lookup = slack.channels.list_(exclude_archived=True)
 
     if live:
         result = await exec(lookup)
-        assert isinstance(result[0], slack.Channel)
+        assert isinstance(result.content[0], slack.Channel)
 
     query = iter(lookup)
 
@@ -37,7 +38,7 @@ async def test_channel_list(exec):
     outcome = sendreturn(query, snug.Response(200, CHANNEL_LIST_RESULT))
     assert isinstance(outcome[0], slack.Channel)
     assert len(outcome[0].members) == 2
-    assert outcome.next_cursor == "dGVhbTpDMUg5UkVTR0w="
+    assert outcome.next_query.cursor == "dGVhbTpDMUg5UkVTR0w="
 
 
 @pytest.mark.asyncio