+++ snug/compat.py
@@ -3,6 +3,7 @@ import sys
 from operator import attrgetter
 
 PY3 = sys.version_info > (3, )
+HAS_PEP492 = sys.version_info > (3, 5, 2)
 
 
 if PY3:
@@ -16,6 +17,7 @@ if PY3:
 
     def func_from_method(method):
         return method
+
 else:  # pragma: no cover
     import urllib2 as urllib_request  # noqa
     from singledispatch import singledispatch  # noqa