+++ snug/query.py
@@ -1,5 +1,6 @@
 """Types and functionality relating to queries"""
 import typing as t
+import warnings
 from functools import partial
 
 from .clients import send
@@ -110,7 +111,7 @@ class Query(t.Generic[T]):
         """
         raise NotImplementedError()
 
-    def __execute__(self, client, authenticate):
+    def __execute__(self, client, auth):
         """Default execution logic for a query,
         which uses the query's :meth:`~Query.__iter__`.
         May be overriden for full control of query execution,
@@ -127,7 +128,7 @@ class Query(t.Generic[T]):
         ----------
         client
             the client instance passed to :func:`execute`
-        authenticate: ~typing.Callable[[Request], Request]
+        auth: ~typing.Callable[[Request], Request]
             a callable to authenticate a :class:`~snug.http.Request`
 
         Returns
@@ -138,14 +139,14 @@ class Query(t.Generic[T]):
         gen = iter(self)
         request = next(gen)
         while True:
-            response = send(client, authenticate(request))
+            response = send(client, auth(request))
             try:
                 request = gen.send(response)
             except StopIteration as e:
                 return e.args[0]
 
     # this method is overwritten when importing the _async module (py3 only)
-    def __execute_async__(self, client, authenticate):
+    def __execute_async__(self, client, auth):
         raise NotImplementedError('python 3+ required to execute async')
 
 
@@ -180,52 +181,90 @@ class related(object):
         return self._cls if obj is None else partial(self._cls, obj)
 
 
+# a temporary routine until `auth_method` is removed in version 1.3
+def _make_auth(auth, auth_method):
+    if auth is None:
+        return _identity
+    elif auth_method is None:
+        return auth if callable(auth) else basic_auth(auth)
+    else:
+        warnings.warn('The `auth_method` parameter will be removed'
+                      'in version 1.3. Pass a callable to `auth` instead',
+                      DeprecationWarning)
+        return partial(auth_method, auth)
+
+
 def execute(query, auth=None, client=urllib_request.build_opener(),
-            auth_method=basic_auth):
+            auth_method=None):
     """Execute a query, returning its result
 
     Parameters
     ----------
     query: Query[T]
-        the query to resolve
-    auth: T_auth
-        the authentication credentials. If using the default ``auth_method``,
-        ``auth`` must be a (username, password)-tuple.
+        The query to resolve
+    auth: ~typing.Tuple[str, str] \
+        or ~typing.Callable[[Request], Request] or None
+        This may be:
+
+        * A (username, password)-tuple for basic authentication
+        * A callable to authenticate requests.
+        * ``None`` (no authentication)
     client
         The HTTP client to use.
         Its type must have been registered
         with :func:`~snug.clients.send`.
         If not given, the built-in :mod:`urllib` module is used.
     auth_method: ~typing.Callable[[T_auth, Request], Request]
-        the authentication method to use
+        The authentication method to use.
+
+        .. deprecated:: 1.2
+           the `auth_method`-parameter
+
+        .. warning::
+
+           The `auth_method` parameter will be removed in version 1.3.
+           Pass a callable to `auth` to implement different
+           authentication methods.
 
     Returns
     -------
     T
         the query result
     """
-    exec_func = getattr(type(query), '__execute__', _default_execute_method)
-    authenticate = _identity if auth is None else partial(auth_method, auth)
-    return exec_func(query, client, authenticate)
+    exec_fn = getattr(type(query), '__execute__', _default_execute_method)
+    return exec_fn(query, client, _make_auth(auth, auth_method))
 
 
-def execute_async(query, auth=None, client=event_loop, auth_method=basic_auth):
+def execute_async(query, auth=None, client=event_loop, auth_method=None):
     """Execute a query asynchronously, returning its result
 
     Parameters
     ----------
     query: Query[T]
-        the query to resolve
-    auth: T_auth
-        the authentication credentials. If using the default ``auth_method``,
-        ``auth`` must be a (username, password)-tuple.
+        The query to resolve
+    auth: ~typing.Tuple[str, str] \
+        or ~typing.Callable[[Request], Request] or None
+        This may be:
+
+        * A (username, password)-tuple for basic authentication
+        * A callable to authenticate requests.
+        * ``None`` (no authentication)
     client
         The HTTP client to use.
         Its type must have been registered
         with :func:`~snug.clients.send_async`.
         If not given, the built-in :mod:`asyncio` module is used.
     auth_method: ~typing.Callable[[T_auth, Request], Request]
-        the authentication method to use
+        The authentication method to use.
+
+        .. deprecated:: 1.2
+           the `auth_method`-parameter
+
+        .. warning::
+
+           The `auth_method` parameter will be removed in version 1.3.
+           Pass a callable to `auth` to implement different
+           authentication methods.
 
     Returns
     -------
@@ -237,10 +276,8 @@ def execute_async(query, auth=None, client=event_loop, auth_method=basic_auth):
     The default client is very rudimentary.
     Consider using a :class:`aiohttp.ClientSession` instance as ``client``.
     """
-    exec_func = getattr(
-        type(query), '__execute_async__', Query.__execute_async__)
-    authenticate = _identity if auth is None else partial(auth_method, auth)
-    return exec_func(query, client, authenticate)
+    exc_fn = getattr(type(query), '__execute_async__', Query.__execute_async__)
+    return exc_fn(query, client, _make_auth(auth, auth_method))
 
 
 def executor(**kwargs):