+++ tests/test_clients.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 import json
 import sys
 
@@ -103,6 +104,42 @@ class TestSendWithAsyncio:
         assert data['args'] == {}
         assert json.loads(data['data']) == {'foo': 4}
         assert data['headers']['User-Agent'] == 'snug/dev'
+
+    def test_nonascii_headers(self, loop):
+        req = snug.Request('GET', 'http://httpbin.org/get',
+                           headers={'X-Foo': 'blå'})
+        response = loop.run_until_complete(snug.send_async(loop, req))
+        assert response == snug.Response(200, mock.ANY, headers=mock.ANY)
+        data = json.loads(response.content.decode())
+        assert data['url'] == 'http://httpbin.org/get'
+        assert data['args'] == {}
+        assert data['headers']['X-Foo'] == 'blå'
+
+    def test_head(self, loop):
+        req = snug.Request('HEAD', 'http://httpbin.org/anything',
+                           headers={'X-Foo': 'foo'})
+        response = loop.run_until_complete(snug.send_async(loop, req))
+        assert response == snug.Response(200, b'', headers=mock.ANY)
+        assert 'Content-Type' in response.headers
+
+    def test_timeout(self, loop):
+        import asyncio
+
+        req = snug.Request('GET', 'http://httpbin.org/delay/2')
+        with pytest.raises(asyncio.TimeoutError):
+            loop.run_until_complete(
+                snug.send_async(loop, req, timeout=.5))
+
+    def test_redirects(self, loop):
+        req = snug.Request('GET', 'http://httpbin.org/redirect/4')
+        response = loop.run_until_complete(snug.send_async(loop, req))
+        assert response == snug.Response(200, mock.ANY, headers=mock.ANY)
+
+    def test_too_many_redirects(self, loop):
+        req = snug.Request('GET', 'http://httpbin.org/redirect/3')
+        response = loop.run_until_complete(
+            snug.send_async(loop, req, max_redirects=1))
+        assert response == snug.Response(302, mock.ANY, headers=mock.ANY)
 
 
 @live