+++ HISTORY.rst
@@ -4,6 +4,15 @@ Release history
 development
 +++++++++++
 
+1.2.0 (2018-03-21)
+++++++++++++++++++
+
+- ``auth`` parameter accepts callables
+- deprecate ``auth_method`` parameter (to remove in version 1.3)
+- paginated queries
+- make ``asyncio`` client more robust
+- added two new recipes
+
 1.1.3 (2018-03-07)
 ++++++++++++++++++